+++ bot.py
@@ -5,6 +5,7 @@ import re
 from discord.ext import commands
 #from poker.discord_control import *
 from poker.pair import *
+from funny import *
 
 bot = commands.Bot(command_prefix= '.')
 voiceBot = None
@@ -182,44 +183,3 @@ async def game(ctx, oper, arg):
     
     with open('games.pkl', 'wb') as output:
         pickle.dump(games, output)
-
-@bot.command()
-async def todo(ctx):
-    embed = discord.Embed(title="Todo",description="japierdole",color=0x00ff00)
-    embed.add_field(name="Krzysia pojebauo",value="bardzo",inline=False)
-    embed.add_field(name="Krzysia pojebauo",value="bardzo mocno",inline=False)
-    await ctx.message.channel.send(embed=embed)
-
-@bot.event
-async def on_message(message):
-    emoji = discord.utils.get(message.guild.emojis, name='OOF')
-    msg = message.content.lower()
-    x = re.search("^[^a-zA-Z0-9]*[o]+of[^a-zA-Z0-9]*$",msg)
-    if x != None:
-        await message.add_reaction(emoji)
-    x = re.search('dzień dobry .*', msg)
-
-    if bot.user.mentioned_in(message):
-        if re.search('dzień dobry .*', msg) != None:
-            await message.channel.send('Dzień dobry! '+message.author.mention)
-        elif re.search('dobranoc .*', msg) != None:
-            await message.channel.send('Dobranoc! '+message.author.mention)
-    await bot.process_commands(message)
-
-    try:
-        for attachments in message.attachments:
-            att = attachments.url
-            for ext in ['.jpg','.png','.jpeg','.PNG','.JPEG','.JPG']:
-                if att.endswith(ext):
-                    for x in ['🇦', '🇧' , '🇨', '🇩', '🇪', '🇫']:
-                        emoji = bot.get_emoji(x)
-                        await message.add_reaction(x)
-    except IndexError:
-        pass
-
-@bot.event
-async def on_reaction_add(reaction,user):
-    for react in reaction.message.reactions:
-        async for users in react.users (limit=None,after=None):
-            if react != reaction and users == user and user != bot.user:
-                await reaction.remove(user)
